+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "Dippet" extension will be documented in this file.
 
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
+## [Unreleased]
+
+- Support multiple commands.
+
 ## [0.1.0]
 
 - Initial release